misc/pytesseract/requirements.txt
opencv_python==4.10.0.84
Pillow==11.0.0
pytesseract==0.3.13
